requirements.txt
numpy
scipy
git+https://github.com/pybox2d/pybox2d
gym[box2d]
tensorflow-gpu==1.4.0
keras==2.1.5
matplotlib
pyglet==1.3.2
h5py
